requirements.txt
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
PyYAML>=5.4.1
matplotlib>=3.4.0
pytest>=6.2.5
black>=21.5b2
flake8>=3.9.0
flask>=2.0.0
plotly>=5.3.0 